ab: Do not fail with summary listing values below noise threshold
In the past, it was possible for a performance A/B-Test to fail with the
message listing all the metrics that had a significant change (p<0.01)
having only entries of changes where the magnitude is below the noise
threshold.

This is because to determine whether we want to print the message, try
to correct for false-positives by considering the average change
reported across all data points for the metric, and if the average of
their relative changes is below the noise threshold, we ignore it.
However, this average includes the metrics not deemed statistically
significant (p>=0.01), so if we have large variation that is not deemed
significant, this can still skew this average, and thus fail the overall
test. Since the failure message only contains significant results, it
will seem like the test failed with data below the noise threshold.

This commit fixes the above by adding a check that ensures that the
average change of a metric across all statistically significant data
points is above the noise threshold before failing the test.

Signed-off-by: Patrick Roy <roypat@amazon.co.uk>