fix: use ext4 rootfs for test_boottime
This bring the setup of this test in line with the
test_boottime_with_net and test_boottime_no_network, where ext4 is used
for better performance.

Long term, we might be able to deduplicate tests here, by eliminating
test_boottime_with_net, as it is subsumed by one of the
parameterizations of test_boottime.

Signed-off-by: Patrick Roy <roypat@amazon.co.uk>