doc: Deduplicate "how to run tests" instructions in tests/README.md
The README contained two sections on how to run tests, one at the top
under the "Running" section, and then again at the end under
"Troubleshooting". Combine these into a single section that contains
the information of both of them.

Also normalize `tools/devtool` and `./tools/devtool` in favor of the
former.

Signed-off-by: Patrick Roy <roypat@amazon.co.uk>