chore: Do not mark test_boottime as nonci
This test is only used for A/B-tests, so from that perspective it does
not matter whether it's marked as nonci or not. However, it is not part
of the nightly metrics pipeline, and the nightly PR pipeline excludes
nonci tests, meaning outside of A/B-tests, this test isnt run at all.
This is still not a functional problem, because the test does not assert
anything. However, it means we do not have regular metrics from this
test. By running it in the nightly PR pipeline, we start collecting
metrics, and once metrics are backfilled, we will be able to remove the
test_bootime_with_network test (while updating our dashboard to display
the metrics from test_boottime instead).

Signed-off-by: Patrick Roy <roypat@amazon.co.uk>