fix: keys in IGNORED dict should be mandatory
Due to the way `is_ignored` was implemented, any dimension inserted into
an entry of the IGNORED directionary became mandatory for all tests that
wished to be A/B-compatible. This commit fixes this (dimension sets that
already do not have the keys that an ignore list entry would require
will simply be considered to not match the entry).

Signed-off-by: Patrick Roy <roypat@amazon.co.uk>