fix: Increase memory pressure in test_deflate_on_oom
Running the test with a 4.14 guest resulting in it failing because a
supposed OOM condition did not result in the balloon being deflated.
However, this is simply because a 4.14 guest kernel is smaller, so more
memory needs to be dirtied for the OOM to be triggered in the first
place.

Signed-off-by: Patrick Roy <roypat@amazon.co.uk>