chore: test to check if killing microvm works as expected
Change:
Add a test to make sure there is no change in behaviour in
killing firecracker_pid.

Reason:
In our CI we pass `--new-pid-ns` along with `--daemonize` and so
firecracker pid can be read from the file `firecracker.pid`
present in the jailers root directory.
However, if there is a change in behaviour that leads to this file not
having the actual firecracker pid then it could cause a regression.

Signed-off-by: Sudan Landge <sudanl@amazon.com>