refactor: remove usage of _screen_firecracker_pid
For the case where Jailer is no daemonized but new_pid_ns
flag is set, we won't be able to fetch Firecracker pid from
the screen process and the right way to get Firecracker PID
is by calling Microvm's self.firecracker_pid so,
remove usage _screen_firecracker_pid.

The change to kill screen process if screen_pid is valid is done to
make the next commit more readable.

Signed-off-by: Sudan Landge <sudanl@amazon.com>