vhost-net: introduce alternative vhost backend for virtio/net device
This patch adds a new backend to Net devices. It can be enabled with
'vhost' bool on the network interface config, like this:

    "network-interfaces": [
        {
            "iface_id": "eth0",
            "host_dev_name": "tap0",
            "vhost": true
        }
    ],

Vhost backend opens host kernel /dev/vhost-net interface, and performs
a setup dance to setup the vhost device with the relevant tap
interface. The effect is that all of the data plane goes directly
between host kernel and the guest, skipping the firecracker VMM.
This drastically reduces the packet latency and increases throughput,
especially in a high-pps scenarios. For example UDP and TCP without
offloads.

The control plane does go through firecracker, due to MMIO limitations.
The exception is interrupt from host to guest, technically they should
go through firecracker VMM, but this is avoidable
by splicing the host eventfd into the guest interruptfd, and
force-returning VIRTIO_MMIO_INT_VRING in the relevant virtio register.
This is the same trick block-vhost-user device uses.

There are couple of missing features:

 - persist (no blockers, just work)
 - mmds (no obvious way to do it, perhaps possible with ebpf)
 - rate_limiting (no obvious way to implement it, perhaps with ebpf)
 - tap/vhost feature negotiation

On the latter point, it would be nice to negotiate some more advanced
tap/vhost features, like USO (UDP segmentation offload), TCP offloads
(flag needed if guest wants to use XDP), VIRTIO_NET_F_MRG_RXBUF (this
might be useful for performance, but benchmarks needed first). Right
now there is no way to express these toggles in the net config, but
this can be done in the future.