fix: do not limit pipeline_perf.py to nonci tests in metrics mode
When running A/B-Tests, the pipeline generated by pipeline_perf.py does
not care about pytest markers, it will run all tests specified and do
the A/B-comparision. However, when running the pipeline in metrics
collection mode, it passes `-m nonci` to pytest, which causes it to
filter out tests that do not have the `nonci` marker. Some of our
A/B-Tests (such as memory overhead) indeed do not have this marker,
leading to confusing errors (where everything works in A/B-mode, but if
just collecting metrics pytest fails with "no tests collected"). This
commit makes the behavior consistent, but also triggering non-nonci
tests in metrics collection mode.

Signed-off-by: Patrick Roy <roypat@amazon.co.uk>