chore(docs): deprecate --start-time-cpu-us parameter
The jailer passes an option to Firecracker `--start-time-cpu-us` with
the start time of the CPU timer. Firecracker substracts that from its
own CPU time to calculate how much CPU time it spends during process
startup. The value the jailer passes is always 0, so we can just use the
Firecracker value.

In any case this value was only supposed to be used within the
Firecracker CI and not meant to be used by end customers.

Signed-off-by: Pablo Barbáchano <pablob@amazon.com>