fix(mmds): Update micro-http for empty MMDS path
When querying a MMDS path that has an empty content, the response didn't
contain 'Content-Length' field, resulting in the client waiting for the
connection closed by the server or timeout. To fix this issue,
micro-http made a change to have the field when the response status code
is not 1XX or 204. This commit updates Cargo.lock file to consume this
change and also adds an integration test to ensure that it doesn't
timeout for an empty MMDS path.

The micro-http's change was made on
https://github.com/firecracker-microvm/micro-http/pull/48.

Signed-off-by: Takahiro Itazuri <itazur@amazon.com>