acpi: add VMGenID support for x86_64
Add support for the Virtual Machine Generation ID (VMGenID) device that
allows notifying the guest about snapshot resume events. The device
itself allocates one page of guest memory and an IRQ line. It stores a
16 bytes cryptographically random number (generation ID) at the
beginning of this page. Once the microVM resumes from a snapshot, it
writes a new 16-byte generation ID and sends an interrupt to the device.

Also add support for the Generic Event Device, an ACPI device which
handles IRQ lines allocated to ACPI devices and routes them as
ACPI notifications to the devices the belong to.

Signed-off-by: Babis Chalios <bchalios@amazon.es>