refactor: Remove function pointer from MMIODevManagerConstructorArgs
The function pointer is called for every device being restored, however
all use-sites of it pointed to the same function
(VmResources::update_from_restored_device), so the
indirection/abstraction is not needed. Removing this function point
removes restrictions on the signature of the function, and thus makes
adding error handling easier (for example, we can now update
update_from_restored_device to validate whether the devices read from
the snapshot are actually consistent with the rest of the vm
configuration).

The entire `SharedDeviceType` enum can probably be removed too, but lets
keep the changes minimal in this commit, as its for the 1.7 release
branch.

Signed-off-by: Patrick Roy <roypat@amazon.co.uk>