test: allow using 6.1 kernel in selected tests
At the moment, we filter out 6.1 guest kernel in all tests apart from
PTP on Graviton. Add a new filter that allows select any kernel that
exists in our CI artifacts folder and use it to create a pytest fixture
for guest kernel 6.1, which works independently of the platform we run
on.

Signed-off-by: Babis Chalios <bchalios@amazon.es>