fix: Have `./tools/ab_test.py analyze` actualy perform an A/B-Test
When invoking the ab_test.py script with two manually generated
test-report.json files, it would compute all the regression and
p-values, but not actually print out which ones are considered failures.
Fix this by moving the assertion to `analyze_data` from
`ab_performance_test`, so that it is also run in `analyze` mode (instead
of only in `run` mode).

Signed-off-by: Patrick Roy <roypat@amazon.co.uk>