fix(test): wait for microVM to boot before snapshotting
In test_vulnerabilities.py we have various tests that check whether a
condition holds after resuming from a snapshot. These checks seem to
consistently fail if we take a snapshot before letting the guest kernel
boot.

Introduce an ssh command to ensure that the guest has booted before
taking the snapshot so that we avoid the issue.

Signed-off-by: Babis Chalios <bchalios@amazon.es>