chore(cpu-template-helper): Add mock kernels
Currently cpu-template-helper requires some resources (kernel, rootfs
and Firecracker config files) to dump a guest CPU config or a
fingerprint or verify a CPU template, although they don't affect the
output essentially. This is because it utilizes
`build_microvm_for_boot()` function imported from vmm crate that
requires those resources. Adding mock kernels is preliminary task to
remove the requirement.

`build_microvm_for_boot()` requires a valid ELF file for x86_64 and a
file containing a valid kernel header for aarch64. So we use a minimal C
program for x86_64 and just write a kernel header to a file for aarch64.

Signed-off-by: Takahiro Itazuri <itazur@amazon.com>