refactor(cpu-template-helper): Change --config optional
Now users no longer need to prepare kernel image file, rootfs file and
Firecracker config file that don't affect the output essentially. If
`--config` is not provided, the CPU template helper tool does the
preparation instead. The parameter itself is preserved for backward
compatibility and special usecases (like changing the number of vCPUs).
In addition, it becomes able to simplify tests a lot.

Signed-off-by: Takahiro Itazuri <itazur@amazon.com>