test: Do not use --config in integration tests
Now that there is no need to specify `--config` to
`cpu-template-helper`, removes `--config` from integration tests and
simplifies them. Also regathers fingerprint files. All the fingerprint
changes are caused by the difference of the number of vCPUs: the
Firecracker's default value is 1 vCPU but the test config value is 2
vCPUs.

Signed-off-by: Takahiro Itazuri <itazur@amazon.com>