refactor(vmm): Exclude SYS_CNTPCT_EL0 from guest CPU config dump
On kernel 6.5, a new timer register SYS_CNTPCT_EL0 becomes available and
it was added to an exception list in integration tests. But from users'
perspective, it is quite harmful for users to need to filter it out on
their side because it isn't useful in CPU template lifecycle and it is
ignoreable. So it should be excluded on the tool side.

Signed-off-by: Takahiro Itazuri <itazur@amazon.com>