fix(vmm): only use memfd if no vhost-user-blk devices configured
Page faults are more expensive for shared memory mapping (like memfd).
For this reason, we only back guest memory with a memfd
if a vhost-user-blk device is configured in the VM, otherwise
we fall back to anonymous private memory.

This is recovering performance demonstrated before commit
027a9929d62c77fa580eaac4792a9acc1bd3ae1c had been merged.

The vhost-user-blk branch in VM creation is not currently
covered by Rust integration tests, because we are looking for
a converging solution, so guest memory is always created
in the same way, and implementing required integration test
infrastructure code (spawing a backend process) would not
be useful long term.

Signed-off-by: Nikita Kalyazin <kalyazin@amazon.com>