deprecate `--start-time-[cpu-]us` flags
These flags are used by the jailer to pass the value that should be
subtracted from the (CPU) time, when emitting the `start_time_us` and
`start_time_cpu_us` metrics. They were never meant to be used by end
customers.

We cannot print a deprecation message, as the jailer unconditionally
passes these two flags to Firecracker. This means that the deprecation
message would be printed unconditionally for all users of the jailer.
Not passing these flags from the jailer to Firecracker is also not an
option, because it would result in a change of behavior (e.g. currently
the metrics are only emitted if the flags are passed, so this would need
to change to have the metrics be _always_ emitted).

Signed-off-by: Patrick Roy <roypat@amazon.co.uk>
Co-authored-by: Pablo Barbáchano <pablob@amazon.com>