fix(test/balloon): remove racy checks
Those negative checks depend on how slow the balloon device activates.
If it activates fast, the checks fail.

One of the checks is already done in unit tests.

The other check we don't have an equivalent for, but the use case should
be rare enough to not matter (configuring a balloon, starting the VM and
checking that replacing it fails). We continute testing that the balloon
can be replaced after it is activated.

Signed-off-by: Pablo Barbáchano <pablob@amazon.com>