refactor(vmm): Remove dumpable MSR indices from undumpable list
`UNDUMPABLE_MSR_RANGES` had two classes of MSR indices: one for really
undumpable MSRs due to disablement of some features and the other for
MSRs that are dumpable but not reasonable to be included in the CPU
template helper tool's fingerprint. The second class is removed from
`UNDUMPABLE_MSR_RANGES`, but instead will be excluded on the CPU
template helper tool's side in the next commit. Thanks to this
separation, the dumped CPU configuration might be usable for other
usecases than CPU template helper, although there isn't such a usecase
right now.

Signed-off-by: Takahiro Itazuri <itazur@amazon.com>