ci: hardcode guest kernel version 5.10
Starting from 5.10.211 the 5.10 guest kernels don't build neither for
x86_64 nor Aarch64 with our kernel configs. There is a missing header
file in arch/x86/kernel/paravirt.c, so when we build with
CONFIG_PARAVIRT on, the build fails.

Pin the 5.10 guest kernel version to the latest one working for us
(5.10.210) until this issue is fixed upstream.

Signed-off-by: Babis Chalios <bchalios@amazon.es>