fix(tests): strip debug information before a testrun
Before df112e39c the tests compiled Firecracked and stripped the
debuginfo. When we moved to running release.sh the behavior changed
since we only strip when making a release.

Solve it by always stripping symbols when building Firecracker in
--release profile.

Some care was taken so it is safe to re-run, so we don't lose the debug
symbols if the original binary has not changed.

Fixes: df112e39ccd558c51494b0d0e4ac8b288f6d140d

Signed-off-by: Pablo Barbáchano <pablob@amazon.com>