chore(vmm): Make IA32_TSX_CTRL MSR serializable
IA32_TSX_CTRL MSR came to be returned by KVM_GET_MSR_INDEX_LIST API
since kernel 6.4. To make IA32_TSX_CTRL MSR restorable from snapshot,
adds it to a list of serializable MSRs. Prior to kernel 6.4, Firecracker
doesn't include the MSR in snapshot, but it is confirmed that the effect
of the MSR is valid even after snapshot restore as long as the host
applies the right mitigation.

Signed-off-by: Takahiro Itazuri <itazur@amazon.com>