tests: microvm: refactor Jailer class
- make the jailer base class generic
- move `screen` bits to a special purpose "jailer"

Abstract the idea of the jailer into a Runner. The idea is that we can
create different runner modes for different cases. For example we in
this commit we have the jailer and screen+jailer, but we could extend it
to others, like not using the Firecracker jailer at all.

Signed-off-by: Pablo Barbáchano <pablob@amazon.com>