tests: drop pandas dependency
Rewrite the only test that needs pandas.

pandas is a heavy dependency that weighs in around 50MB. It also
recently started printing a warning that in the future it will require
pyarrow, which is another 40 MB.

Pandas is great, however our usage of it is minimal.

Size of devctr does not change much, but we avoid the warning, and not
having to deal with pyarrow in the future.

Example output:
```
MSR removed 0x13 before=0x0
MSR changed 0x179 before=0x20ffff after=0x20
MSR added 0x17 after=0x0
MSR added 0x11 after=0x25ba008
```

Signed-off-by: Pablo Barbáchano <pablob@amazon.com>