test(balloon): add delay in stats update test
It looks like there is a race condition leading to intermittent
failures.  The test does the following:
 - request pages from the driver
 - change the polling interval to trigger stats update
 - read stats and check that available memory has updated
Processing of the inflation request may take time due to memory
allocation happening in the guest.

The change inserts a 1 second delay to increase the chance of
memory allocation completion by the time of the check.

Signed-off-by: Nikita Kalyazin <kalyazin@amazon.com>