use correctly ftrace event for determining guest page faults in ARM
The ftrace event emitted when a guest causes a page fault on ARM is
called `kvm_guest_fault`, not `guest_page_fault`. I actually had this
correctly on my ARM testing instance, but apparently I messed it up when
it came to actually writing this commit on my laptop.

Keep the metric we emit from the test named `guest_page_faults` though,
to avoid needing to change code that expects this name.

Fixes: e44f6adb9c2725bc6d5982f9e6db5f3df7035272
Signed-off-by: Patrick Roy <roypat@amazon.co.uk>