refactor: Add utility to wait for the guest coming up
Many tests attempts to run a command inside a microVM via SSH connection
to check if the guest is up and running or to wait for microvm to boot.
Commands used to check were diverse: `true`, `sync` or `ls`. To unify
this process, add a new method `wait_for_up()`.

Signed-off-by: Takahiro Itazuri <itazur@amazon.com>