test: Change ssh's log level from quiet to error
We passed `-q` option to `ssh` command in order to exclude warning
messages from stderr, because some tests expect stderr is empty in the
remote command's result. However, that made it difficult to debug errors
of `ssh` command itself. Change the log level from quiet mode to error
mode that displays more information on failure but doesn't output
anything to stderr on success.

Signed-off-by: Takahiro Itazuri <itazur@amazon.com>