test: Increase timeout for microvm to boot up
It was observed that "Guest-boot-time" was not found in Firecracker log
within 1 second in the nightly test. Set timeout to 5 seconds that is
big enough for the guest to boot up even if it run concurrently on the
same host. Otherwise, the boot time metric is always reported less than
1 second and lead to a false sense. By setting enough timeout, in the
next occurrence of timeout error, it can be told that something went
wrong inside guests. In happy cases, it will exit within 1 second, so it
usually doesn't make the test running for long time.

Signed-off-by: Takahiro Itazuri <itazur@amazon.com>