fix: power off vcpus only during startup
Vcpus should be in power off state during kernel boot,
but during snapshot restore we do not need to power them off
even thought kvm does not complain. To do this we stop applying
the `KVM_ARM_VCPU_POWER_OFF` to the `kvi` during restore.
Also if we need to store `kvi` in the state, do this before adding
`KVM_ARM_VCPU_POWER_OFF` to it, so that during restore `kvi`
will not have `KVM_ARM_VCPU_POWER_OFF` set.

Signed-off-by: Egor Lazarchuk <yegorlz@amazon.co.uk>