feat(aarch64): store kvi all the time
Previously we were storing kvi for vcpus only
when cpu template was used. This was done to
save vcpu configuration for snapshot restoration
step, because when vcpus are restored, we don't
have access to the cpu template to apply changes
again. But when cpu template was not used we
did not store already created kvi and were
getting new one from KVM. This is a redundant work
which we can avoid if we store kvi all the time.
Also storing kvi unconditionally simplifies
the code significantly.

Signed-off-by: Egor Lazarchuk <yegorlz@amazon.co.uk>