test: stop all monitors before killing microvm
If a test fails, the FCmetricsMonitor keeps running in some tests. After
we kill the microvm and remove the resources, the monitor complains it
cannot find the metrics file.

This is tricky to fix simply since we don't have a handle to the monitor
either at fixture teardown or in the Microvm class.

Add a hook in the Microvm class to register monitors, so we can stop all
of them before killing the microvm.

Also adapt existing monitors to fix this pattern.

Signed-off-by: Pablo Barbáchano <pablob@amazon.com>