test: use a different private IP block
Some other software or cloud providers can make use of private IP space,
which can result in a conflict. Ideally we would detect what block to
use, but just changing the block is much simpler and quicker.

Use the last /16 within the 10.x.x.x /8 to try to avoid these kind of
conflicts.

Signed-off-by: Pablo Barbáchano <pablob@amazon.com>