ci: encapsulate Buildkite pipeline into a class
Consolidate all the pipeline functionality that was repeated in each
pipeline script into a class.

The result is more succinct and readable, and enables adding features in
a single place.

While porting existing scripts to the pipeline, some of them gain new
features like pipeline_cpu_template and pipeline_cross, which now accept
the same arguments as the rest.

Signed-off-by: Pablo Barbáchano <pablob@amazon.com>