chore(ci): double pytest timeout duration temporarily
Currently our A/B testing builds Firecracker for each test and each
environment while running tests. As the Firecracker build time
increases, some tests timeout the current duration (300s). To mitigate
this issue temporarily, sets it to 600s. It should be set it back to the
original value after fixing the issue in PR #4624.

Signed-off-by: Takahiro Itazuri <itazur@amazon.com>