refactor(test): Have `restore_from_snapshot` return snapshot object
Make `Microvm.restore_from_snapshot` return a `Snapshot` object. The
return value describes the snapshot files inside the microvm's jail,
which are potentially a copy of the given `Snapshot`.

This will be used so that tests that restore snapshots in a loop will
be able to clean them up again easily.

Signed-off-by: Patrick Roy <roypat@amazon.co.uk>