fix: Delete snapshot files in test_snapshot_ab.py
The `test_restore_latency` test resumes 30 microVMs in a loop. Each
iteration copies the snapshot that is to be restored into the respective
microvm's jail, where they are not cleaned up until the end of the test,
e.g. after all 30 iterations have completed. This can cause a problem if
any of these 30 iterations fail. In this case, we copy all files in the
chroot of each of the 30 microvms into `test_results` (even the ones
that completed successfully). This can result in us running out of disk
space if it happens with large snapshots.

To avoid this, delete the copies of the snapshots at the end of their
respective iteration.

Signed-off-by: Patrick Roy <roypat@amazon.co.uk>