fix(test_rng): do not test for virtio-rng performance
Currently, our test for virtio-rng rate limiter tests if the performance
obtained by the device is within 10% of the rate limiter limit. This
does not only test whether the rate limiter works (achieved bandwidth >
10% of limit), but also if the device performs at that same limit.

This test is supposed to be a functional test, not a performance one. In
fact, it does not run in isolation. Moreover, we have not thoroughly
profiled the performance we can achieve from the device, so we cannot be
sure that the limits in the functional test correspond to the
capabilities of the device. If we want, we should do that as part of a
performance test.

This commit removes the check for the achieved bandwidth being up to 10%
less than the limit.

Signed-off-by: Babis Chalios <bchalios@amazon.es>