test_5_snapshots: remove SIGSTOP/SIGCONT pair
The test creates, and resumes from 5 consecutively taken snapshots. This
means it involves 6 microVMs: The seed microVM that is originally
booted, and then 5 restored microVMs. The SIGSTOP/SIGCONT logic was
supposed to ensure that the in-guest state of the vsock `socat` process
would be consistent across snapshot-restore boundaries: We "stop" the
process before taking a snapshot, and "continue" it after restoring.
However, there are two problems with our implementation:
- No `SIGSTOP` was sent before taking the first snapshot, and
- The SIGSTOP/SIGCONT signals related to later snapshots were sent to
  the wrong microVM (they were sent inside the original, booted microVM,
  but never inside the restored microVMs, since the `vm` variable from
  outside the `for` loop was reused, instead of the `microvm` variable
  from inside the loop).

Together, these mean that we never actually stopped/continued any
`socat` servers. Thus drop the whole logic related to it, since it is
nonsense. Instead, replace with with a `time.sleep(2)`, as seemingly the
ssh connection to the wrong microVM was inducing exactly the right delay
(measured at slightly above 1s, which I rounded up to 2) to reduce
connection related intermittent failures (which is why we thought the
SIGSTOP/SIGCONT pair was fixing the intermittent failures).

Signed-off-by: Patrick Roy <roypat@amazon.co.uk>