test: log stdout/stderr in case of command timeouts
In case the `timeout` argument is passed to `utils.run_cmd_sync`, and a
timeout actually happens, `proc.communicate` will raise a
`TimeoutException`. This means that any output the command might have
produced before timing out will be lost. Thus, catch the
`TimeoutException` instead and log stdout and stderr before re-raising
it, as described by the python documentation [1].

We log at `warning` level instead of error so that the logs are only
printed if the tests fails, as there are a few tests
(test_deflate_on_oom for example) that `expect` the `TimeoutExpired`
exceptions to ignore them (in which case logging them anyway would be
noise).

[1]: https://docs.python.org/3/library/subprocess.html#subprocess.Popen.communicate

Signed-off-by: Patrick Roy <roypat@amazon.co.uk>