refactor(test): Replace ssh.run with ssh.check_output where possible
Replace all occurences of ssh.check_output that were immediately
followed by an assertion on the return code being zero with a call to
`ssh.check_output` instead. Apart from reducing lines of code, this has
the advantage that we unify various assertion message formats (going
from no message at all to various combinations of stderr and stdout)
into a single format that include the command run, its stdout and
stderr, and its return code.

The instances replaced in this commit are based on grepping the code
base for `assert \w+ == 0`. I make no claims of this catching
everything. I also did not replace assertions that instead of checking
the status code, assert `stderr` to be empty. I assume most of these can
also be dropped, but I did not want to inspect every single command we
execute to check whether output to stderr always implies a non-zero
return code. So they remain as they are.

No functional change intended.

Signed-off-by: Patrick Roy <roypat@amazon.co.uk>