refactor(test): do not print gitlint output twice
`test_gitlint` was doing `assert False` inside an exception handler,
which caused python to think "exception occured during exception
handling", meaning it would print both the assertion failure, as well as
the original exception. However, the assertion message contained the
message of the original exception, so we just ended up printing
everything twice.

Use `ignore_return_code=True` to avoid the exception and just assert on
the return code ourselves, to avoid this.

Signed-off-by: Patrick Roy <roypat@amazon.co.uk>