refactor(tests): Rename `utils.run_cmd` to `utils.check_output`
Similar to python's `subprocess` module and our own `SSHConnection`
class, it'd be nice if in `utils` we'd have two functions.
`check_output`, which raises an exception if the return code of an
executed command is non-zero, and `run_cmd`, which by default does not
check the return code (but has a parameter to enable this).

The easiest way to migrate towards that is by renaming the existing
function to `check_output` (since that's what its current semantic is,
and what its callers expect), and then in a follow up commit
reintroduce `run_cmd` as a new function. Then all callers of
`check_output` that set `ignore_return_code=True` can be migrated, and
`ignore_return_code` can be dropped from `check_output`.

This commit does the first step towards that migration, renaming
`run_cmd` to `check_output`. No functional change intended.

Signed-off-by: Patrick Roy <roypat@amazon.co.uk>