refactor: update writing to descriptor logic
When packet could not be written to the descriptor
chain due to some error, we were still marking it as used,
even though we could reuse it later. Alse we were running
this logic in the loop that will perform same actions and
encounter same error multiple times for no reason.

Now we don't loop and do writing to the descriptor chain
only once. If writing is not success, we don't consume
descriptor.

Signed-off-by: Egor Lazarchuk <yegorlz@amazon.co.uk>