test: Use os.kill instead of check_output("kill -9 ... || true")
There's library methods for killing processes in python, so let's just
use those instead of shelling out.

Also move the `os.kill` call into the try-except block, to get logging
in case Firecracker unexpected died before we could kill it.

The only change in behavior is that we no longer silently ignore the
case where we fail to kill Firecracker because it is already dead.

Signed-off-by: Patrick Roy <roypat@amazon.co.uk>