test: increase retries in wait_process_termination
In `test_signals.py::test_sigxfsz_handler`, we seem to need to wait on
average 6 seconds. Prior to unifying all the "wait for FC to die" tests,
this one just had an endless loop, and the retry parameters passed to
`wait_process_termination` which it is using now instead do not provide
sufficient waiting. So simply increase the waiting in
`wait_process_termination` to 10s.

Signed-off-by: Patrick Roy <roypat@amazon.co.uk>