fix(test): Stop guest from panicking in boottime tests on AMD
In our boottime tests, we use a custom `init` script to signal to
Firecracker (via MMIO) that booting has finished. This script then
`execve`s into `/sbin/init` (e.g. systemd) to continue the normal boot
process. However, when execve-ing, we forgot to NULL-terminate the
`envp` array, resulting a buffer over-run that caused systemd to crash
on m6a.metal instances.

Thus, NULL-terminate the array to fix the buffer-overrun. For this
change to take effect, we need to rebuild CI artifacts.

Signed-off-by: Patrick Roy <roypat@amazon.co.uk>