fix(ci): Do not prevent tests from running if style check fails
Since the "Style" step is before the "wait" step, everything after
"wait" will only get executed if the Style check passes (in addition to
the build steps). This is undesirable, as often temporary testing
commits are used (which are not intended to get merged, and thus often
fail the style check).

Fix this by moving the style step after the "wait", but by giving it a
`depends_on: null` entry, which instructs buildkite to executed it
immediately (e.g. without waiting) anyway, without blocking the wait
step [1]

[1]: https://buildkite.com/docs/pipelines/dependencies#explicit-dependencies-in-uploaded-steps

Signed-off-by: Patrick Roy <roypat@amazon.co.uk>