Refactor Bus and BusDevice
Previously, Bus instances stored Arc<Mutex<BusDevices>>, this has now been
updated so that the Bus only stores BusDevices, and the device wrapped
inside the BusDevice enum is now wrapped in an ArcMutex instead. This
should allow devices to be shared, as they are no longer owned
exclusively by the MMIODeviceManager.

Signed-off-by: James Curtis <jxcurtis@amazon.co.uk>