fix(test): remove `wait_for_up` from `test_vmgenid`
The `wait_for_up()` sends a dummy SSH command with a timeout of 10s,
with the intention of blocking the integration test until a VM is
responsive after booting. However, post-snapshot restore it runs into
issue #4099, where on AMD inside the guest, sshd gets stuck inside a
nanosleep syscall for longer than intended (and thus longer than the 10s
timeout).

Alternatively, we could pass `lapic=notscdeadline` in the kernel
commandline arguments, but we do not have an easy way to _append_ to the
default kernel command line from integration tests, and I'd like to
avoid duplicating our default across a bunch of places.

Signed-off-by: Patrick Roy <roypat@amazon.co.uk>