test: eliminate `Microvm.expect_kill_by_signal`
All tests that formerly set this property now use `mark_killed`. The
only other location that set this value was `kill()` itself, to avoid
errors when kill() ends up being called twice. However, this is already
managed by the `Microvm._killed` attribute, so this guarding is not
necessary. Thus, remove the `expect_kill_by_signal`, as it is useless.

Signed-off-by: Patrick Roy <roypat@amazon.co.uk>